Split the endpoint module
