+++ apiron/endpoint/endpoint.py
@@ -1,4 +1,3 @@
-import collections
 import logging
 import warnings
 
@@ -128,75 +127,3 @@ class Endpoint:
 
     def __str__(self):
         return self.path
-
-
-class JsonEndpoint(Endpoint):
-    """
-    An endpoint that returns :mimetype:`application/json`
-    """
-
-    def __init__(self, *args, path='/', default_method='GET', default_params=None, required_params=None, preserve_order=False):
-        super().__init__(path=path, default_method=default_method, default_params=default_params, required_params=required_params)
-        self.preserve_order = preserve_order
-
-    def format_response(self, response):
-        """
-        Extracts JSON data from the response
-
-        :param requests.Response response:
-            The original response from :mod:`requests`
-        :return:
-            The response's JSON content
-        :rtype:
-            :class:`dict` if ``preserve_order`` is ``False``
-        :rtype:
-            :class:`collections.OrderedDict` if ``preserve_order`` is ``True``
-        """
-
-        return response.json(object_pairs_hook=collections.OrderedDict if self.preserve_order else None)
-
-    @property
-    def required_headers(self):
-        return {'Accept': 'application/json'}
-
-
-class StreamingEndpoint(Endpoint):
-    """
-    An endpoint that streams data incrementally
-    """
-
-    streaming = True
-
-    def format_response(self, response):
-        """
-        Stream response in chunks
-
-        :param requests.Response response:
-            The original response from :mod:`requests`
-        :return:
-            The response's content
-        :rtype:
-            generator
-        """
-
-        return response.iter_content(chunk_size=None)
-
-
-class StubEndpoint:
-    """
-    A stub endpoint designed to return a pre-baked response
-
-    The intent is to allow for a service to be implemented
-    before the endpoint is complete.
-    """
-
-    def __init__(self, stub_response=None, **kwargs):
-        """
-        :param stub_response:
-            A pre-baked response, like 'stub response' or {'stub': 'response'}
-        :param **kwargs:
-            Arbitrary parameters that can match the intended real endpoint.
-            These don't do anything for the stub but streamline the interface.
-        """
-        self.endpoint_params = kwargs if kwargs else {}
-        self.stub_response = stub_response or 'stub for {}'.format(self.endpoint_params)
