+++ apiron/endpoint/streaming_endpoint.py
@@ -1,0 +1,23 @@
+from apiron.endpoint.endpoint import Endpoint
+
+
+class StreamingEndpoint(Endpoint):
+    """
+    An endpoint that streams data incrementally
+    """
+
+    streaming = True
+
+    def format_response(self, response):
+        """
+        Stream response in chunks
+
+        :param requests.Response response:
+            The original response from :mod:`requests`
+        :return:
+            The response's content
+        :rtype:
+            generator
+        """
+
+        return response.iter_content(chunk_size=None)
